% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{learning.checkQuality}
\alias{learning.checkQuality}
\title{Check a Learning Quality Value}
\usage{
learning.checkQuality(quality)
}
\arguments{
\item{quality}{the quality value}
}
\value{
\code{TRUE} if the quality value is permissible, \code{FALSE} otherwise
}
\description{
Check whether a value \code{quality} is a permissible value for
  the \code{quality} field of \code{\link{learning.Result}}.
}
