% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{learning.checkSize}
\alias{learning.checkSize}
\title{Check a Learning Size Value}
\usage{
learning.checkSize(size)
}
\arguments{
\item{size}{the size value}
}
\value{
\code{TRUE} if the size  value is permissible, \code{FALSE} otherwise
}
\description{
Check whether a value \code{size} is a permissible value for the
  \code{size} field of \code{\link{learning.Result}}.
}
