% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\docType{class}
\name{learning.Result-class}
\alias{learning.Result-class}
\alias{learning.Result}
\title{The Result of the Application of a Machine Learning or Optimization Technique}
\description{
This class is a base class for capturing the result of the
  application of a Machine Learning or optimization method. Actually, this
  base class only stores a quality measure value, which must be a finite
  positive real number or zero and a measure of the "size" of the solution,
  e.g., the number of parameters that have been determined, which must be a
  positive integer number or zero. Sub-classes can then add more information,
  such as the data of the actual solution.
}
\section{Slots}{

\describe{
\item{\code{quality}}{the quality of the result, smaller values are better}

\item{\code{size}}{the size of the result, e.g., the number of fixed parameters}
}}

