% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\name{learning.Result.new}
\alias{learning.Result.new}
\title{Create a New Instance of \code{\link{learning.Result}}.}
\usage{
learning.Result.new(quality, size = .Machine$integer.max)
}
\arguments{
\item{quality}{the quality of the result, smaller values are better}

\item{size}{the size of the result, e.g., the number of fixed parameters}
}
\value{
the new instance
}
\description{
Create a New Instance of \code{\link{learning.Result}}.
}
