% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{rank.by.comparison}
\alias{rank.by.comparison}
\title{Rank a Given Data Set based on a Comparator Function}
\usage{
rank.by.comparison(data, comparator)
}
\arguments{
\item{data}{the data set, a list or vector of data elements}

\item{comparator}{a binary comparator function}
}
\value{
a vector of ranks
}
\description{
Rank the elements in the set \code{data} based on binary
  comparisons via a \code{comparator} function. The function
  \code{comparator} accepts two parameters, \code{a} and \code{b}, which will
  be two elements from \code{data}. The function then returns a negative if
  the first element (\code{a}) should receive a lower rank than the second
  element (\code{b}). It returns a positive value if \code{b} should be
  ranked smaller than \code{a}. It in all other cases (i.e., if \code{0} or
  \code{NaN} are returned), neither element is preferred.
}
