% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\docType{methods}
\name{learning.Result.finalize}
\alias{learning.Result.finalize}
\alias{learning.Result.finalize,learning.Result-method}
\title{Finalize a Learning Result}
\arguments{
\item{object}{the instance of \code{\link{learning.Result}}}

\item{object}{the \code{\link{learning.Result}} instance}
}
\value{
the finalized \code{\link{learning.Result}} instance

the same as \code{object}
}
\description{
This method can be invoked after a model learning process has
  completely finished. It should canonicalize the result as far as possible
  and may perform other tasks. It will not change any characteristics of the
  result, though. Since this work may be time-consuming, we put it in an
  extra method. Assume, for instance, that your learning task is to fit many
  models to a data set and then select a specific model for future use. You
  would invoke \code{learning.Result.finalize} only on this selected model.

This default implementation does nothing and returns the instance as-is
}
